+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use a full text parser for backlinks search in their short form
   (/username/shortId). It prevents false positives, when some URL in text
   contains a shortlink-looking fragment.
+- Keep the 'invitation_id' field during the user's personal data deletion to
+  preserve connection with the inviting user.
 
 ## [2.18.1] - 2024-01-26
 ### Fixed